TranscribeClient/requirements-dbclient.txt
openai-whisper
paho-mqtt
torch
soundstream
lightning
torchmetrics
wandb